signer: remove explicitly mentioned script type
The signing scripts will include the appropriate
hash-bang (#!) and executable permissions, allowing
them to be run directly. These permissions are
preserved when the scripts are copied into the
rpmbuild sandbox environment.

To support signing with the Gerrit
pipeline—which will use an ELF executable
version of the signer—this change is required.

Change-Id: I236347aab7011f34a1da258e7eb12ac173bc6c68
Signed-off-by: Kuntal Nayak <kuntal.nayak@broadcom.com>
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/25827
Tested-by: Michelle Wang <michelle.wang@broadcom.com>
Reviewed-by: Bo Gan <bo.gan@broadcom.com>