yajl: fix CVE-2022-24795
As mentioned in
https://github.com/brianmario/yajl-ruby/security/advisories/GHSA-jj47-x69x-mxrm,
this is likely not practically exploitable on 64-bit systems.

However, let's patch it anyways just to be safe. There is some debate
upstream on the best fix here, and yajl has not been maintained for
some 10+ years now (and thus no security fixes there either).

Applied the same fix as currently upstream in Fedora.

Change-Id: Ic3877df33c06dc3d5dbe4d6308c3077351b3a893
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/25889
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>
Reviewed-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>
Reviewed-by: Harinadh Dommaraju <harinadh.dommaraju@broadcom.com>