photon-lic-tool: Enable manual review of certain files
Some files can be erroneously labeled with incorrect
licenses by scancode - one example is for AGPL,
which is detected in some cases where the closer match
should be regular GPL.

To handle such cases, add a field in the config yaml
for manual review of these files. The correct license
can be specified there.

The list of manually reviewed files is considered
per archive.

Since there can be multiple subdirectories under BUILD/,
and there is no way for us to tell which is the final build
directory, we may need to specify two paths:
1. archive relative path
2. final build directory path

2 is only necessary if it differs from the archive relative path.
For an example, see jitterentropy* files in Linux. These are copied
from the unpacked jitterentropy archive into the final Linux build
directory.

Both paths are necessary to delete ALL copies of this file from BUILD/,
before the scan is started.

rsync is done as an example.

Change-Id: I91e36a493b7f28d1255dbf1786ed5c8e724efea4
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/25945
Reviewed-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>
Reviewed-by: Alexey Makhalov <alexey.makhalov@broadcom.com>
Reviewed-by: gerrit-photon <svc.photon-ci@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>
Tested-by: Alexey Makhalov <alexey.makhalov@broadcom.com>