photon-lic-tool: add option to scan all sources in config.yaml
Adds a new option '--config_yaml <path to new config.yaml>',
which will allow to scan all sources in the given config.yaml
and output the new config.yaml to the path passed to `--config_yaml`.

This allows for use of manual review data when parsing the source
tarball licenses. However, because some files are manually reviewed
as "" in order to ignore licenses for files which are not packaged,
e.g some doc files, this will end up falsely ignoring licenses
from the source tarball and that is not ok.

Leaving the logic there, in future it may be ok to enable it.

For '--config_yaml', '--yaml' refers to a directory, where
Scancode yaml output for each source directory will
be placed.

Also updated license_tree to alphabetically sort all sub-expressions
inside parantheses. This is to help produce more consistent
expressions.

Note: Use ruamel.yaml to preserve comments in the config.yaml

Note2: Formatted with black

Change-Id: Ifcae2a9f642cd05eccbbe6e25605154737ece0ac
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26193
Reviewed-by: Michelle Wang <michelle.wang@broadcom.com>
Reviewed-by: gerrit-photon <svc.photon-ci@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>
Reviewed-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26264
Reviewed-by: <mukul.sikka@broadcom.com>
Tested-by: <mukul.sikka@broadcom.com>
Reviewed-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>