photon-lic-tool: better handle cache removes
Improve handling for cases where file paths are incorrect,
or file is missing, when setting up the cache.

Improved by checking for file existence and using absolute
paths to avoid ambiguity.

Also, corner case discovered where a file under systemd
has '\' in the name. Scancode output converts the backslash
to a forward slash, and this causes a file not found error.

Fix this by checking for the file existence. In this corner
case, we will just not cache the result for this file.

Change-Id: Ia943a2fa83d87f6bba81ece53867d0520ed906f5
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26206
Reviewed-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>
Reviewed-by: gerrit-photon <svc.photon-ci@broadcom.com>
Reviewed-by: Michelle Wang <michelle.wang@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>