linux: disable external IVs for gcm(aes)
External IVs are not allowed for gcm(aes) - block them entirely.
As seqiv seems to be the only IV generator used for gcm(aes)
for IPSec, block anything except for seqiv.

This is achieved by marking the aead_request with the magic
number (randomly set on first access). Validation checks that
the internal_iv field matches the magic number.

Magic random number helps to block OOT kernel modules from attempting to set
the internal iv field in the aead_request to circumvent the check,
as they will not have access to the random magic number.

Testing:
        1. rfc4106(gcm_base(ctr(aes-aesni),ghash-generic)) - BLOCKED
        2. (__rfc4106-gcm-aesni-avx) (__rfc4106(gcm(aes)) - BLOCKED
        3. seqiv(rfc4106(gcm(aes))) - SUCCESSFUL
        4. seqiv(rfc4106(gcm_base(ctr(aes-aesni),ghash-generic))) - SUCCESSFUL

Also, allocate requests with kzalloc to ensure internal_iv is init to 0.

Change-Id: I688a8ff8885387771ec2eb811c5a16d9b0192cac
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26328
Reviewed-by: Ankit Jain <ankit-aj.jain@broadcom.com>
Reviewed-by: Alexey Makhalov <alexey.makhalov@broadcom.com>
Tested-by: Alexey Makhalov <alexey.makhalov@broadcom.com>
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>