linux-fips-canister: packages cansiter inside a tarball
commit c2994349ceb6 ("linux-fips-canister: package plain files instead
of a tarball") eliminated tarball and packages plain fips_canister.o.
But it introduced aregression. rpmbuild stripped this object file into
nonconsumable form.
Move canister object file back to the tarball.

It is not a full revert. .cmd file still not needed and directory structure
inside the tarball can be simplified.

Change-Id: I50cb494569ecd57f22a88b8c0b6a17e31d582c14
Signed-off-by: Alexey Makhalov <alexey.makhalov@broadcom.com>
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26405
Reviewed-by: Vamsi Krishna Brahmajosyula <vamsi-krishna.brahmajosyula@broadcom.com>
Tested-by: Vamsi Krishna Brahmajosyula <vamsi-krishna.brahmajosyula@broadcom.com>
Reviewed-by: Ankit Jain <ankit-aj.jain@broadcom.com>