photon-lic-tool: improve memory consumption
From experimentation, it was discovered that the `--timeout` option
causes scancode to use massive quantities of memory when processing
multiple large files simultaneously. For example, when scanning
several ~50MB files with two threads, each thread was observed
consuming upwards of 9GB of memory, while without the timeout option,
the same scanning consumed only ~1GB of memory each.

The time to scan also seems to increase with a longer timeout, for some
reason.

To avoid this annoyance, and avoid any OOM issues, we can switch
the timeout to the subprocess module which handles the scancode
subprocess.

Default is set to no timeout, to allow external timeouts for entire
photon-lic-tool command. But can be specified with --scancode_timeout.

Also, add a function to attempt to dynamically calculate a safe
number of threads which can be used in parallel. This is based
on observed memory usage on my machine. This is only used
in cases where the thread count isn't specified. This will attempt
to use as many threads as possible while using a safe amount of memory.
For running multiple photon-lic-tool in parallel, it is better to
manually limit the CPUs to avoid conflicts.

Change-Id: I84c3d97b1da83adb8436fc907fdd8d5dcf1e79d9
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26559
Reviewed-by: Michelle Wang <michelle.wang@broadcom.com>
Reviewed-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>
Tested-by: Michelle Wang <michelle.wang@broadcom.com>