photon-lic-tool: make manual review entries more flexible
Because we can now have multiple different paths to the same
file depending on --config_yaml or --build_spec options,
we should make the file path search flexible.

If we find the file path in at least one place, we don't need
to throw an error - assume the other file path(s) don't exist
for this scan type.

If we don't find the file at any of the given paths, then
we can throw an error.

Change-Id: I7c606fb915ab90f55d271e1e19584a9908429a47
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26649
Tested-by: gerrit-photon <svc.photon-ci@broadcom.com>
Reviewed-by: gerrit-photon <svc.photon-ci@broadcom.com>
Reviewed-by: Shreenidhi Shedi <shreenidhi.shedi@broadcom.com>
Reviewed-by: Brennan Lamoreaux <brennan.lamoreaux@broadcom.com>