package-builder: Move tdnf snapshot repos to rpmPath
When creating local snapshots for tdnf sandbox it is
assumed that both rpmPath and snapshot repoPath are
in same filesystem. This is not true when a overlay
is used to represent stage/RPMS.

Use symlinks instead of hardlinks and move the tdnf
repo base path under stage/RPMS to handle this.

Suggested-by: Bo Gan <bo.gan@broadcom.com>
Signed-off-by: Vamsi Krishna Brahmajosyula <vamsi-krishna.brahmajosyula@broadcom.com>

Change-Id: I5e4a026929f059e7338624b2b4f52cafff8b0818
Reviewed-on: http://photon-gerrit.lvn.broadcom.net/c/photon/+/26840
Reviewed-by: Bo Gan <bo.gan@broadcom.com>
Tested-by: Michelle Wang <michelle.wang@broadcom.com>
Reviewed-by: Michelle Wang <michelle.wang@broadcom.com>